+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased Changes
+
+### Features
+
+- Adds the `unwrap_bump!` macro, which extracts the bump from the instruction `Context` ([#36](https://github.com/saber-hq/vipers/pull/36)).
+
 ## [2.0.0] - 2022-02-21
 
 ### Features